backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
google-cloud-storage
requests
google-auth
